pyproject.toml
[tool.poetry]
name = "aritstmanagement"
version = "0.1.0"
description = ""
authors = ["cloco-prajwol <shakya.prajwol@cloco.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^25.1"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
